Split slices
